feat(citas): Añadir campo "servicios", valor decimal y selector de fechas
- Ampliar la interfaz de Cita con el campo "servicios"
- Cambiar "ServicioCita.valor" de BigInt a Decimal para mayor precisión monetaria
- Añadir el componente "DisabledDatePicker" con fechas bloqueadas desde la API
- Corregir el formato de fecha en las tablas de administración para gestionar correctamente la hora UTC
- Habilitar la creación de citas desde la página de administración del cliente
- Mejorar la consistencia del diseño de la interfaz de usuario en los formularios
- Almacenar la sesión del cliente en cookies durante el proceso de registro

+++ prisma/migrations/20260225041113_init/migration.sql
@@ -159,7 +159,7 @@ CREATE TABLE "ServicioCita" (
     "id" TEXT NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "servicio" TEXT,
-    "valor" BIGINT,
+    "valor" DECIMAL(65,30),
     "usuarioId" TEXT,
     "citaId" TEXT,
